refactor: create a function for rendering a gallery page
create a function for rendering a whole gallery page

MODULE-1-PART-3-gallery
